Move serialization and deserialization routines
