+++ db/filters/base.py
@@ -139,66 +139,12 @@ class And(MultiParameter, Branch, Predicate):
 class Or(MultiParameter, Branch, Predicate):
     type: BranchPredicateType = static(BranchPredicateType.OR)
 
-def getSAFilterSpecFromPredicate(pred: Predicate) -> dict:
-    if isinstance(pred, Leaf):
-        if isinstance(pred, SingleParameter):
-            return {'column': pred.column, 'op': pred.saId(), 'value': pred.parameter}
-        elif isinstance(pred, MultiParameter):
-            return {'column': pred.column, 'op': pred.saId(), 'value': pred.parameters}
-        elif isinstance(pred, NoParameter):
-            return {'column': pred.column, 'op': pred.saId()}
-        else:
-            raise Exception("This should never happen.")
-    elif isinstance(pred, Branch):
-        if isinstance(pred, SingleParameter):
-            subject = getSAFilterSpecFromPredicate(pred.parameter)
-            return {pred.saId(): [subject]}
-        elif isinstance(pred, MultiParameter):
-            subjects = [ getSAFilterSpecFromPredicate(subject) for subject in pred.parameters ]
-            return {pred.saId(): subjects}
-        else:
-            raise Exception("This should never happen.")
-    else:
-        raise Exception("This should never happen.")
-
+# TODO rename to getPredicateSubClassByTypeStr
 def getPredicateSubClassByType(predicateTypeStr: str) -> Union[Type[LeafPredicateType], Type[BranchPredicateType]]:
     for subClass in allPredicateSubClasses:
         if subClass.type.value == predicateTypeStr:
             return subClass
     raise Exception(f'Unknown predicate type: {predicateTypeStr}')
-
-def getPredicateFromMAFilterSpec(spec: dict) -> Predicate:
-    def getFirstDictKey(dict: dict) -> Any:
-        return next(iter(dict))
-    try:
-        assert isinstance(spec, dict)
-        predicateTypeStr = getFirstDictKey(spec)
-        predicateSubClass = getPredicateSubClassByType(predicateTypeStr)
-        predicateBody = spec[predicateTypeStr]
-        if issubclass(predicateSubClass, Leaf):
-            columnName = predicateBody['column']
-            if issubclass(predicateSubClass, SingleParameter):
-                return predicateSubClass(column=columnName, parameter=predicateBody['parameter'])
-            elif issubclass(predicateSubClass, MultiParameter):
-                return predicateSubClass(column=columnName, parameters=predicateBody['parameters'])
-            elif issubclass(predicateSubClass, NoParameter):
-                return predicateSubClass(column=columnName)
-            else:
-                raise Exception("This should never happen.")
-        elif issubclass(predicateSubClass, Branch):
-            if issubclass(predicateSubClass, SingleParameter):
-                parameterPredicate = getPredicateFromMAFilterSpec(predicateBody)
-                return predicateSubClass(parameter=parameterPredicate)
-            elif issubclass(predicateSubClass, MultiParameter):
-                parameterPredicates = \
-                    [ getPredicateFromMAFilterSpec(parameter) for parameter in predicateBody ]
-                return predicateSubClass(parameters=parameterPredicates)
-            else:
-                raise Exception("This should never happen.")
-        else:
-            raise Exception("This should never happen.")
-    except:
-        raise BadFilterFormat("Parsing of Mathesar filter specification failed.")
 
 
 class BadFilterFormat(SABadFilterFormat):
